#TBSFMD-177:customer and organizer registration foriegn key issue resolved

+++ src/main/java/com/uol/smqa/dtos/response/LoginResponseDTO.java
@@ -22,5 +22,5 @@ public class LoginResponseDTO {
 
     private String token;
 
-    private Long expiresIn;
+    private Integer expiresIn;
 }
